moved asymmetry stuff to new file
